added analytics events
